requirements.txt
# Core ML Libraries
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.36.2
diffusers==0.30.2
accelerate==0.25.0
safetensors==0.4.1

# Image/Video Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
imageio==2.33.1
imageio-ffmpeg==0.4.9

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
numpy==1.26.3
scipy==1.11.4
requests==2.31.0